Apply review updates.
- Add to deprecation list
- Add to derecation index
- Add transformation example
- Correct spelling
- Remove base class
