+++ deprecation-index.yml
@@ -192,6 +192,11 @@
   PHPStan: 'Call to deprecated function entity_load(). Deprecated in drupal:8.0.0 and is removed from drupal:9.0.0. Use the entity type storage''s load() method.'
   Examples:
     - entity_load.php
+'entity_view()':
+  Rector: EntityViewRector.php
+  PHPStan: 'Use the entity view builder''s view() method for creating a render array'
+  Examples:
+    - entity_view.php
 'node_load()':
   Rector: NodeLoadRector.php
   PHPStan: 'Call to deprecated function node_load(). Deprecated in drupal:8.0.0 and is removed from drupal:9.0.0. Use Drupal\node\Entity\Node::load().'
