+++ src/Rector/Deprecation/EntityViewRector.php
@@ -2,23 +2,36 @@
 
 namespace DrupalRector\Rector\Deprecation;
 
-use DrupalRector\Rector\Deprecation\Base\EntityViewBase;
 use Rector\Core\RectorDefinition\CodeSample;
 use Rector\Core\RectorDefinition\RectorDefinition;
+use Rector\Core\Rector\AbstractRector;
+use PhpParser\Node;
+
 
 /**
- * Replaced deprecated entity_load() calls.
+ * Replaced deprecated entity_view() calls.
  *
- * See https://www.drupal.org/node/2266845 for change record.
+ * See https://www.drupal.org/node/3033656 for change record.
  *
  * What is covered:
- * - See EntityLoadBase.php
+ * - Static replacement
+ * - The reset parameter is excluded.
  *
  * Improvement opportunities
- * - See EntityLoadBase.php
+ * - Include support for cache rest parameter.
  */
-final class EntityViewRector extends EntityViewBase
+final class EntityViewRector extends AbstractRector
 {
+
+    /**
+     * @inheritdoc
+     */
+    public function getNodeTypes(): array
+    {
+        return [
+            Node\Expr\FuncCall::class,
+        ];
+    }
 
     /**
      * @inheritdoc
@@ -37,6 +50,46 @@ $rendered = \Drupal::entityTypeManager()->getViewBuilder($entity
 CODE_AFTER
             )
         ]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function refactor(Node $node): ?Node
+    {
+        if ($this->getName($node->name) !== 'entity_view') {
+            return NULL;
+        }
+
+        $name = new Node\Name\FullyQualified('Drupal');
+
+        $entityTypManager = new Node\Identifier('entityTypeManager');
+
+        $var = new Node\Expr\StaticCall($name, $entityTypManager);
+
+        $getViewBuilder_method_name = new Node\Identifier('getViewBuilder');
+
+        $entity_reference = $node->args[0]->value;
+        $getEntityTypeId_method_name = new Node\Identifier('getEntityTypeId');
+
+        $entityRef_type_id = new Node\Expr\MethodCall($entity_reference, $getEntityTypeId_method_name);
+
+        $view_builder = new Node\Expr\MethodCall($var, $getViewBuilder_method_name, [$entityRef_type_id]);
+
+        $view_method_name = new Node\Identifier('view');
+
+        $view_args = [
+            $node->args[0]->value,
+            $node->args[1]->value,
+        ];
+
+        if (isset($node->args[2])) {
+            $view_args[] = $node->args[2]->value;
+        }
+
+        $view = new Node\Expr\MethodCall($view_builder, $view_method_name, $view_args);
+
+        return $view;
     }
 
 }
